refactor(complete): Split out bash dynamic code
